Importer - use the CAT instead of the VULN as the template's :data
In some instances you'd want to know the port affected by the problem,
and this information is at the CAT level. In the previous version of the
plugin the information would be lost as we only focused on the nested
VULN element.

This commit makes the core source of information the CAT/VULN element to
have the best of both worlds.

To make this work with the template_service and dradis-plugins there is
some trickery we had to do. A normal XML might look like this:

CAT
 \ VULN1
 \ VULN2

But the template_service is expecting to deal with a single item at a
time. This is why for every VULN we duplicate the structure, but prune
all the other VULN siblings and pass a CAT with exactly 1 nested VULN to
the template_service.